requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
pinecone
docx2txt
